wget -c for continue download if file exists
I've had to run the notebook twice every time it's from a fresh notebook. (Because it always requires me to kernel restart after the first wget/update/etc.) This results in double-downloading the .zstd file. (The new one becomes a .1 file.) The -c parameter should allow wget to skip the second download if the file already exists. (Or it will finish downloading if it's not completely downloaded.)